requirements.txt
alembic==1.9.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.9
Boruta==0.3
catalogue==1.0.2
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorlover==0.3.0
comm==0.1.2
contourpy==1.0.6
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
databricks-cli==0.17.4
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docker==6.0.1
entrypoints==0.4
executing==1.2.0
fastapi==0.89.1
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
fqdn==1.5.1
funcy==1.17
future==0.18.2
gensim==3.8.3
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
htmlmin==0.1.12
hydra-core==1.3.1
identify==2.5.13
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.7.0
importlib-metadata==5.2.0
importlib-resources==5.10.2
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kmodes==0.12.2
lightgbm==3.3.4
llvmlite==0.37.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mlflow==2.1.1
mlxtend==0.19.0
multimethod==1.9.1
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
nodeenv==1.7.0
notebook==6.5.2
notebook_shim==0.2.2
numba==0.54.1
numexpr==2.8.4
numpy==1.20.3
oauthlib==3.2.2
omegaconf==2.3.0
packaging==22.0
pandas==1.5.2
pandas-profiling==3.6.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==2.6.2
plotly==5.11.0
pre-commit==2.21.0
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pycaret==2.3.10
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
PyJWT==2.6.0
pyLDAvis==3.3.1
pynndescent==0.5.8
pyod==1.0.7
pypandoc==1.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.0
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.12.2
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==2.3.9
SQLAlchemy==1.4.46
sqlparse==0.4.3
srsly==1.0.6
stack-data==0.6.2
starlette==0.22.0
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
terminado==0.17.1
textblob==0.17.1
thinc==7.4.6
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typeguard==2.13.3
typing_extensions==4.4.0
umap-learn==0.5.3
uri-template==1.2.0
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.17.1
visions==0.7.5
wasabi==0.10.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wordcloud==1.8.2.2
yellowbrick==1.2.1
zipp==3.11.0
